task-queue/requirements.txt
celery
redis
slackclient==2.1.0